refactor: Convert InstanceRefs to package structure
This commit refactors the InstanceRefs module from a single file to a
proper package structure, following AUTOSAR M2 meta-model organization
and improving code maintainability.

Changes:
- Convert Components/InstanceRefs.py to Components/InstanceRefs/__init__.py
- Add missing imports in DataElements/__init__.py for
  VariableInAtomicSWCTypeInstanceRef and ParameterInAtomicSWCTypeInstanceRef
- Update parser imports to use new package structure
- Update writer to work with new package structure
- Update all affected tests to use new import paths

Fixes flake8 F821 errors for undefined names in type annotations.

Closes #364
