+++ src/armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/Components/InstanceRefs/__init__.py
@@ -187,5 +187,82 @@ class InnerPortGroupInCompositionInstanceRef(AtpInstanceRef):
         return self
 
 
+class OperationInAtomicSwcInstanceRef(AtpInstanceRef, ABC):
+    def __init__(self):
+        if type(self) == OperationInAtomicSwcInstanceRef:
+            raise TypeError("OperationInAtomicSwcInstanceRef is an abstract class.")
+        
+        super().__init__()
+
+        self.baseRef: RefType = None
+        self.contextPortRef: RefType = None
+        self.targetOperationRef: RefType = None
+
+    def getBaseRef(self):
+        return self.baseRef
+
+    def setBaseRef(self, value):
+        self.baseRef = value
+        return self
+
+    def getContextPortRef(self):
+        return self.contextPortRef
+
+    def setContextPortRef(self, value):
+        self.contextPortRef = value
+        return self
+
+    def getTargetOperationRef(self):
+        return self.targetOperationRef
+
+    def setTargetOperationRef(self, value):
+        self.targetOperationRef = value
+        return self
+
+
+class POperationInAtomicSwcInstanceRef(OperationInAtomicSwcInstanceRef):
+    def __init__(self):
+        super().__init__()
+
+        self.contextPPortRef: RefType = None
+        self.targetProvidedOperationRef: RefType = None
+
+    def getContextPPortRef(self):
+        return self.contextPPortRef
+
+    def setContextPPortRef(self, value):
+        self.contextPPortRef = value
+        return self
+
+    def getTargetProvidedOperationRef(self):
+        return self.targetProvidedOperationRef
+
+    def setTargetProvidedOperationRef(self, value):
+        self.targetProvidedOperationRef = value
+        return self
+
+
+class ROperationInAtomicSwcInstanceRef(OperationInAtomicSwcInstanceRef):
+    def __init__(self):
+        super().__init__()
+
+        self.contextRPortRef: RefType = None
+        self.targetRequiredOperationRef: RefType = None
+
+    def getContextRPortRef(self):
+        return self.contextRPortRef
+
+    def setContextRPortRef(self, value):
+        self.contextRPortRef = value
+        return self
+
+    def getTargetRequiredOperationRef(self):
+        return self.targetRequiredOperationRef
+
+    def setTargetRequiredOperationRef(self, value):
+        self.targetRequiredOperationRef = value
+        return self
+
+
